Done itemsUI

+++ Assets/Scripts/Api/Models/Item.cs
@@ -9,7 +9,7 @@ namespace ExordiumGamesAssignment.Scripts.Api.Models
         public string name;
         public string description;
         public string image_url;
-        public decimal price;
+        public double price;
         public int retailer_id;
         public int item_category_id;
     }
